Refactor to use new data structure for tic-tac-toe games.
This is a new boolean data structure.
We now have two matrices indicating if a mark was placed or not.
This has the advantages of a smaller memory footprint.
It has the disadvantages of making creating data a little harder.

There are some immediate time improvements.
The following was the result from running search.
1.79
1.77
1.87
1.82
1.86
That's down around 25% from the previous times.